Move FileSystemSnapshotBuilder to core
It isn't used in `snapshots` at all.
