+++ subprojects/core/src/test/groovy/org/gradle/internal/fingerprint/impl/FileSystemSnapshotBuilderTest.groovy
@@ -14,13 +14,19 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.snapshot
+package org.gradle.internal.fingerprint.impl
 
 import org.gradle.api.internal.cache.StringInterner
 import org.gradle.internal.file.FileType
 import org.gradle.internal.file.impl.DefaultFileMetadata
 import org.gradle.internal.hash.FileHasher
 import org.gradle.internal.hash.HashCode
+import org.gradle.internal.snapshot.DirectorySnapshot
+import org.gradle.internal.snapshot.FileSystemLocationSnapshot
+import org.gradle.internal.snapshot.FileSystemSnapshot
+import org.gradle.internal.snapshot.FileSystemSnapshotVisitor
+import org.gradle.internal.snapshot.RegularFileSnapshot
+import org.gradle.internal.snapshot.RelativePathSegmentsTracker
 import org.gradle.util.TextUtil
 import spock.lang.Specification
